requirements.txt
pymongo==3.12.1
python-dateutil==2.8.2
rich==10.15.1
PyYAML==6.0
fabric==2.6.0